requirements.txt
python-dotenv~=1.0.1
langchain~=0.3.12
langgraph~=0.2.45
openai>=1.30.0
pymupdf
faiss-cpu
sentence-transformers
numpy~=1.23.0
tqdm~=4.67.1
requests~=2.32.4
beautifulsoup4~=4.12.3
networkx~=3.2.1
pydantic~=2.11.7
streamlit~=1.12.0
orjson~=3.10.10
altair<5
